Modification to enforce the datatype within the mode_aggregator method and add an assertion to the unit tests to show that the output dtype matches the input dtype.

+++ improver/categorical/modal_code.py
@@ -270,7 +270,7 @@ class ModalCategory(BaseModalCategory):
         mode_result[counts < minimum_significant_count] = (
             self.code_max - self.unset_code_indicator
         )
-        return self.code_max - np.squeeze(mode_result)
+        return (self.code_max - np.squeeze(mode_result)).astype(data.dtype)
 
     @staticmethod
     def _set_blended_times(cube: Cube) -> None:
@@ -320,7 +320,6 @@ class ModalCategory(BaseModalCategory):
             result = cube
         else:
             result = cube.collapsed("time", self.aggregator_instance)
-            result.data = result.data.astype(cube.dtype)
         self._set_blended_times(result)
 
         result = self._prepare_result_cube(
